requirements.txt
requests>=2.31.0
tomli>=2.0.0  # For TOML parsing
urllib3>=2.0.0  # For retry/backoff in requests
python-dotenv>=1.0.0  # For .env file support
semgrep>=1.0.0  # For static code analysis
setuptools<81.0.0  # Avoid pkg_resources deprecation in newer versions
dependency-check

# AI Agent dependencies
openai>=1.0.0  # For OpenAI GPT-4 integration
anthropic>=0.18.0  # For Claude integration
psutil>=5.9.0  # For system metrics collection

# Note: Ensure these are installed in your environment:
# - git
# - npm (for Node.js projects)
# - go (for Go projects)
# - ruby (for Ruby projects)
# - java (for Java projects)
safety
pip-audit
# Note: The following tools need to be installed separately:
# - npm audit (comes with Node.js)
# - govulncheck (go install golang.org/x/vuln/cmd/govulncheck@latest)
# - bundle-audit (gem install bundler-audit)
# - OWASP Dependency-Check (https://owasp.org/www-project-dependency-check/)


# API Dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-multipart>=0.0.6
diagrams>=0.23.0
